+++ packages/apollo-language-server/src/providers/schema/engine.ts
@@ -12,6 +12,7 @@ import {
 } from "./base";
 
 import { GetSchemaByTag } from "../../graphqlTypes";
+import { Debug } from "../../utilities";
 
 export class EngineSchemaProvider implements GraphQLSchemaProvider {
   private schema?: GraphQLSchema;
@@ -83,6 +84,13 @@ export class EngineSchemaProvider implements GraphQLSchemaProvider {
   ): SchemaChangeUnsubscribeHandler {
     throw new Error("Polling of Engine not implemented yet");
     return () => {};
+  }
+
+  async resolveFederatedServiceSDL() {
+    Debug.error(
+      "Cannot resolve a federated service's SDL from engine. Use an endpoint or a file instead"
+    );
+    return;
   }
 }